+++ src/components/link-preview/embedly.jsx
@@ -1,76 +1,65 @@
-/* global embedly */
-import PropTypes from 'prop-types';
-import React from 'react';
-import { connect } from 'react-redux';
+import React, { memo, useEffect, useRef } from 'react';
+import { useSelector } from 'react-redux';
 
 import { darkTheme } from '../select-utils';
-import ScrollSafe from './scroll-helpers/scroll-safe';
 import * as heightCache from './scroll-helpers/size-cache';
 
-class EmbedlyPreview extends React.Component {
-  static propTypes = { url: PropTypes.string.isRequired };
+export default memo(function EmbedlyPreview({ url }) {
+  const box = useRef();
+  const feedIsLoading = useSelector((state) => state.routeLoadingState);
+  const isDarkTheme = useSelector(darkTheme);
 
-  link = null;
-  state = { updCounter: 0 };
+  useEffect(() => {
+    initEmbedly();
+    box.current.innerHTML = `<a
+      href="${url.replace(/"/g, '&quot;')}",
+      data-card-controls="0"
+      data-card-width="400px"
+      data-card-recommend="0"
+      data-card-align="left"
+      data-card-theme="${isDarkTheme ? 'dark' : 'light'}"
+    ></a>`;
+    window.embedly('card', box.current.firstChild);
+  }, [feedIsLoading, isDarkTheme, url]);
 
-  setLink = (el) => (this.link = el);
+  return (
+    <div
+      ref={box}
+      className="embedly-preview link-preview-content"
+      data-url={url}
+      style={{ height: `${heightCache.get(url, 0)}px` }}
+    />
+  );
+});
 
-  componentDidMount() {
-    embedly('card', this.link);
-  }
-
-  componentDidUpdate() {
-    embedly('card', this.link);
-  }
-
-  UNSAFE_componentWillReceiveProps(nextProps) {
-    if (this.props.feedIsLoading && !nextProps.feedIsLoading) {
-      this.setState({ updCounter: this.state.updCounter + 1 });
-    }
-  }
-
-  render() {
-    // We use 'key' to force re-render HTML-code on
-    // the feed update or on the props.url change
-    return (
-      <div
-        key={`${this.props.url}##${this.state.updCounter}`}
-        className="embedly-preview link-preview-content"
-        data-url={this.props.url}
-        style={{ height: `${heightCache.get(this.props.url, 0)}px` }}
-      >
-        <a
-          ref={this.setLink}
-          href={this.props.url}
-          data-card-controls="0"
-          data-card-width="400px"
-          data-card-recommend="0"
-          data-card-align="left"
-          data-card-theme={this.props.darkTheme ? 'dark' : 'light'}
-        />
-      </div>
-    );
-  }
-}
-
-function select(state) {
-  return {
-    feedIsLoading: state.routeLoadingState,
-    darkTheme: darkTheme(state),
-  };
-}
-
-export default ScrollSafe(connect(select)(EmbedlyPreview));
-
-embedly('on', 'card.resize', (iframe) => {
-  let cont = iframe.parentNode;
-  while (cont && !cont.classList.contains('embedly-preview')) {
-    cont = cont.parentNode;
-  }
-  if (!cont) {
+let embedlyInitialized = false;
+function initEmbedly() {
+  if (embedlyInitialized) {
     return;
   }
-  const height = iframe.offsetHeight;
-  cont.style.height = `${height}px`;
-  heightCache.set(cont.dataset.url, height);
-});
+  embedlyInitialized = true;
+
+  const id = 'embedly-platform';
+  if (document.getElementById(id)) {
+    return;
+  }
+  window.embedly =
+    window.embedly || ((...args) => (window.embedly.q = window.embedly.q || []).push(args));
+  const script = document.createElement('script');
+  script.id = id;
+  script.async = true;
+  script.src = `//cdn.embedly.com/widgets/platform.js`;
+  const [lastScript] = document.getElementsByTagName('script');
+  lastScript.parentNode.insertBefore(script, lastScript);
+
+  // Listen for resize events
+  window.embedly('on', 'card.resize', (iframe) => {
+    const cont = iframe.closest('.embedly-preview');
+    if (!cont) {
+      return;
+    }
+    const height = iframe.offsetHeight;
+    cont.style.height = `${height}px`;
+    heightCache.set(cont.dataset.url, height);
+  });
+}